add wip api call to progress graph

+++ osc_backend/userinfo/models.py
@@ -14,11 +14,11 @@ class Goal(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     goal_date = models.DateField(auto_now_add=True)
     screen_goal = models.FloatField(default=3)
-    meditation_goal = models.IntegerField(default=10)
+    meditation_goal = models.FloatField(default=10)
     excercise_goal = models.FloatField(default=1)
-    screen_time = models.IntegerField(default=0)
-    meditation_time = models.IntegerField(default=0)
-    excercise_time = models.IntegerField(default=0)
+    screen_time = models.FloatField(default=0)
+    meditation_time = models.FloatField(default=0)
+    excercise_time = models.FloatField(default=0)
 
     def __str__(self):
         return self.goal_date
